ai-ml/think-spark/requirements.txt
streamlit>=1.35
python-dotenv>=1.0
openai>=1.30
anthropic>=0.30
google-generativeai>=0.7
groq>=0.9
mistralai>=1.0
requests>=2.31
